enable CLI to read from live database
